chore(elementArrayFinder): separate elementFinder into its own file
